requirements.txt
numpy==1.21.5
opencv-python==4.5.5.62
PyYAML==3.12
